Module pallet_dex::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

add_liquidity

Add new liquidity to an existing pool (asset_a and asset_b) An equivalent amount of liquidity pool assets will be minted to the user who’s sending liquidity.

create_liquidity_pool

Creates a liquidity pool and also a new asset for it that will be placed in the event Event::PoolCreated.

remove_liquidity

It removes from an existing liquidity pool both assets “a” and “b”. Both amounts are calculated taking into account the amount of the liquidity pool asset that will be burned. Optionally, minimums can be set so that the user receives “at least” some amounts. The new reserves of the pool cannot be zero. Once burned, this emits the event Event::LiquidityRemoved.

swap_exact_in_for_out

It swaps the exact asset amount (exact_in) to a calculated output of the another asset if it’s greater or equal than the minimum output the user expects The output is calculated using the input minus the fee so that its difference with the “real” output remains in the liquidity pool as an incentive for the users

swap_in_for_exact_out

It swaps the exact asset amount (exact_out) to a calculated input of the another asset if it’s lower or equal than the maximum input the user expects The input is calculated using the output minus the fee so that its difference with the “real” input remains in the liquidity pool as an incentive for the users