Function pallet_dex::pallet::dispatchables::add_liquidity

pub fn add_liquidity<T: Config>(
    asset_a_id: <<T as Config>::Fungibles as Inspect<<T as Config>::AccountId>>::AssetId,
    asset_b_id: <<T as Config>::Fungibles as Inspect<<T as Config>::AccountId>>::AssetId,
    asset_a_balance_desired: <<T as Config>::Fungibles as Inspect<<T as Config>::AccountId>>::Balance,
    asset_b_balance_desired: <<T as Config>::Fungibles as Inspect<<T as Config>::AccountId>>::Balance,
    asset_a_balance_min: <<T as Config>::Fungibles as Inspect<<T as Config>::AccountId>>::Balance,
    asset_b_balance_min: <<T as Config>::Fungibles as Inspect<<T as Config>::AccountId>>::Balance
)

Add new liquidity to an existing pool (asset_a and asset_b) An equivalent amount of liquidity pool assets will be minted to the user who’s sending liquidity.

Based on the uniswap contract, <https://github.com/Uniswap/v2-periphery/blob/0335e8f7e1bd1e8d8329fd300aea2ef2f36dd19f/contracts/UniswapV2Router02.sol#L33>, this extrinsic requires “desired” and “min” balances of both assets. 1. Using the asset_a_balance_desired and the reserves it tries to calculate the optimal_b_balance 2. If it’s greater than the user’s asset b amount minimum or desired, it calculates the optimal asset a using again the reserves and the asset b desired amount. Therefore, this function will receive both ranges of amounts and will select the asset_a_balance_desired or asset_b_balance_desired and the other asset balance following the current ratio of the liquidity pool.

Once added, this emits in the event Event::PoolCreated.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::add_liquidity.